requirements.txt
pandas
boto3
psycopg2-binary  # For PostgreSQL
python-dotenv
sqlalchemy
aws-cdk-lib
constructs